fix(ios): 筛选栏标签宽度显式转成 CGFloat
避免 Double 传给 SwiftUI frame 在 Swift 6 下编不过。

+++ codes/mobile/ios/Timia/Features/Schedule/ScheduleMapView.swift
@@ -169,10 +169,10 @@ struct ScheduleMapView: View {
             Text(label)
                 .font(.caption.weight(.medium))
                 .foregroundStyle(.secondary)
-                .frame(width: ScheduleMapFilterBarLayout.labelColumnWidth, alignment: .leading)
+                .frame(width: CGFloat(ScheduleMapFilterBarLayout.labelColumnWidth), alignment: .leading)
             content()
         }
-        .frame(minHeight: ScheduleMapFilterBarLayout.rowMinHeight, alignment: .center)
+        .frame(minHeight: CGFloat(ScheduleMapFilterBarLayout.rowMinHeight), alignment: .center)
     }
 
     private var mapCanvas: some View {
